+++ app_backend/app/config.py
@@ -11,9 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from core.telemetry import FormatType, LogLevel
 from datarobot.core.config import DataRobotAppFrameworkBaseSettings
-
-from app.telemetry import FormatType, LogLevel
 
 
 class Config(DataRobotAppFrameworkBaseSettings):