+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [11.5.4] - 2026-02-03
+
+### Added
+
+- Add error toasts to chat mutation hooks
+- Set a limit for input prompt length by
+
+### Fixed
+
+- Handle 401 response for external users
+- Fix circular symlink between app_backend and core by moving telemetry into core
+
 ## [11.5.3] - 2026-02-26
 
 ### Added